fix term aggregation u32::MAX overflow issue
(cherry picked from commit 1e859fd78d71986a7d53bc096b0fd27bfa91aec2)

+++ src/aggregation/bucket/composite/collector.rs
@@ -264,7 +264,7 @@ impl SegmentCompositeCollector {
             dict.insert(
                 key,
                 IntermediateCompositeBucketEntry {
-                    doc_count: agg.count,
+                    doc_count: agg.count as u64,
                     sub_aggregation: sub_aggregation_res,
                 },
             );
